+++ test/Raven.CodeAnalysis.Tests/Completion/CompletionServiceMemberAccessTests.cs
@@ -214,6 +214,31 @@ string.
     }
 
     [Fact]
+    public void GetCompletions_AfterDot_OnNameOfExpression_ReturnsStringMembers()
+    {
+        var code = """
+struct Main {
+}
+
+func Run() {
+    nameof(Main).
+}
+""";
+
+        var syntaxTree = SyntaxTree.ParseText(code);
+        var compilation = Compilation.Create("test", new CompilationOptions(OutputKind.DynamicallyLinkedLibrary))
+            .AddSyntaxTrees(syntaxTree)
+            .AddReferences(TestMetadataReferences.Default);
+
+        var service = new CompletionService();
+        var position = code.LastIndexOf('.') + 1;
+
+        var items = service.GetCompletions(compilation, syntaxTree, position).ToList();
+
+        Assert.Contains(items, i => i.DisplayText == "Length");
+    }
+
+    [Fact]
     public void GetCompletions_AfterDot_OnEnumType_ReturnsEnumMembers()
     {
         var code = """